+++ server/configs.js
@@ -1,7 +1,7 @@
 import mongoose from "mongoose";
 
 
-// mangoose db connection configuration
+// Mongoose db connection configuration
 export const connectDB = async () => {
     const conn = await mongoose.connect(process.env.DATABASE_URL);
     console.log(`MongoDB connected: ${conn.connection.host}`);