+++ CHANGELOG.md
@@ -1,16 +1,27 @@
 # Changelog
 All changes to this project will be documented in this file
 
+## 2026-08-11
+### Added
+- OTP generation endpoint for login & reset password 
+- Resend OTP cede endpoint for login & reset password
+- OTP code verification endpoints
+
+## 2026-08-10
+### Added
+- Resend OTP cede endpoint for create accout 
+### Fixed
+- Fixing user save to database when unstable network and returning of user not exist when network stable
+
 
 ## 2026-08-06
 ### Added
 - OTP generation flow for users during creating account.
-- OTP code verification endpoint to complete the account creation flow.
+- OTP code verification endpoint to complete the account creation.
 
 ## 026-07-30
 ### Fixed
 - Resolved missing words and typo error in backend error responses.
-
 
 ## 2026-07-29
 ### Added